+++ sdk/azcore/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.1.5 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
 ## 1.1.4 (2022-10-06)
 
 ### Bugs Fixed